+++ src/main/java/com/example/demo/model/Driver.java
@@ -22,12 +22,12 @@ public class Driver extends Audit{
     private String id;
     private String name;
     private String phone;
-    private Long mark;
+    private Integer mark;
     private String licenseNumber;
     @DBRef
     private TaxiOffice taxiOffice;
 
-    public Driver( String name, String phone, long mark, String licenseNumber, TaxiOffice taxiOffice) {
+    public Driver( String name, String phone, int mark, String licenseNumber, TaxiOffice taxiOffice) {
         this.name = name;
         this.phone = phone;
         this.mark = mark;
